Cleanup the temp commit, and test more permutations.
